Fixed scraping errors, analyzed more features and added random forest regression model
Fixed scraping error that was repeatedly happening with one or two songs on the Top 100 playlist, figured out I was parsing incorrectly, which caused my indexing to not grab the correct link to click on and give an error.
Scraped and analyzed some extra features, such as number of promos a song has, lyric length, number of followers an artist has. Want to figure out how I can get BPM and Key of the songs maybe?
Used a random forest regression model, but the MSE is quite high, so I probably need to do more data exploration/feature engineering and possibly switch to another model.